{{cookiecutter.project_name}}/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6

SQLAlchemy==1.4.37
python-dotenv==0.20.0
mysqlclient==2.1.0
pytest==7.1.2
requests==2.27.1
python-multipart==0.0.5
python-jose==3.3.0